`init` with scraping and crawling
I think this might be the full thing... **lol** it wasn't hard to create, didn't think it would be but that **was** faster than I expected! I will have to do some more testing and docs, probably write a short overview in the README. I will see how well this works.